+++ CHANGELOG.md
@@ -12,6 +12,7 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 - When switching match type, set match mode to `No Sets`
 - Drop focus from swap players button after clicked
 - Always show change player order modal on official matches
+- Sort tournament matches by putting played matches on top
 
 #### Fixed
 - Bug where tournaments couldn't be filtered by location